Disable "JOIN" button if no room name is entered. Extract base WelcomePageContainer functionality into an abstract class.
