+++ features/welcome/components/base/index.js
@@ -1,0 +1,1 @@
+export * from './WelcomePageContainer';
